Seperated files into valid and invalid connections
