+++ tests/core/predefined/InvalidConnections.ts
@@ -5,45 +5,6 @@ export const emptyConnections: EngineConnections = {
     output: {}
 }
 
-export const simpleValidConnection: EngineConnections = {
-    input: {
-        "rootINPUT0": {
-            self: {
-                ioId: "rootINPUT0",
-                nodeId: "root",
-                type: CONNECTION_TYPE.INPUT,
-                index: 0,
-            },
-            connections: [
-                {
-                    ioId: "constOneEngineNodeOUTPUT0",
-                    nodeId: "constOneEngineNode",
-                    type: CONNECTION_TYPE.OUTPUT,
-                    index: 0
-                }
-            ]
-        }
-    },
-    output: {
-        "constOneEngineNodeOUTPUT0": {
-            self: {
-                ioId: "constOneEngineNodeOUTPUT0",
-                nodeId: "constOneEngineNode",
-                type: CONNECTION_TYPE.OUTPUT,
-                index: 0
-            },
-            connections: [
-                {
-                    ioId: "rootINPUT0",
-                    nodeId: "root",
-                    type: CONNECTION_TYPE.INPUT,
-                    index: 0,
-                }
-            ]
-        }
-    }
-}
-
 export const missingOutgoingConnection: EngineConnections = {
     input: {
         "rootINPUT0": {
